+++ simple-stack-example-mvp/src/main/java/com/zhuinden/simplestackdemoexamplemvp/presentation/paths/statistics/StatisticsCoordinator.java
@@ -1,8 +1,13 @@
 package com.zhuinden.simplestackdemoexamplemvp.presentation.paths.statistics;
 
+import android.view.View;
+
 import com.zhuinden.simplestackdemoexamplemvp.util.BaseCoordinator;
 
 import javax.inject.Inject;
+
+import butterknife.ButterKnife;
+import butterknife.Unbinder;
 
 /**
  * Created by Owner on 2017. 01. 26..
@@ -12,6 +17,11 @@ public class StatisticsCoordinator
         extends BaseCoordinator<StatisticsView> {
     @Inject
     public StatisticsCoordinator() {
+    }
+
+    @Override
+    protected Unbinder bindViews(View view) {
+        return ButterKnife.bind(this, view);
     }
 
     @Override